Compiled long latency memory.
This memory should be connected to the arbitrer and the cache.
It lasts 3 cycles (from the moment in which the petition bit
is set) until the 32 Bytes data is available at the output,
therefore a cache would have available its data at the beggining of the fifth cycle since the instant of petition=1.