Instruction Cache, Arbiter and Memory work!
We still have to check how they behave together with the processor.
For some reason the memory stays 2 cycles in the 00 stage but
the services are well done.
Even if the memory if the memory is referenced at word level,
we do this only for debugging. Virtually it works at cache_line,
level.